// frontend/src/app/model/object.interface.ts
export interface ObjectI {

    id?: number;
  
    area?: string;
  
    district?: string;
  
    address?: string;
  
    object_address?: string;
  
    object_condition?: string;
  
    object_type?: string;
  
    object_square?: string;
  
    object_length?: string;
  
    units_measurement?: string;
  
    owner?: string;
  
    actual_user?: string;
  
    photo_path?: string;
  
    video_path?: string;
   
    map_path?: string;
   
    area_btn?: string;
  
    spot_area?: string;
  
    height?: string;
   
    wall_material?: string;
   
    year_construction_start?: string;
   
    year_completion_construction?: string;
  
    blocking_services?: string;
  
    checking_department?: string;
   
    number_letter_edo?: string;
  
    date_letter_edo?: string;
  
    comment_object_type?: string;
  
    religion?: string;
   
    okn?: string;
   
    dismantled?: string;
  
    date_dismantling?:string;
   
    person_dismantling?:string;
  
    identified?: string;
  
    oiv?: string;
  
    internal_reconstruction?: string;
  
    manual_status_adjustment?: string;
  
    comment_manual_adjustment?: string;
  
    considered_in?: string;
  
    appendices_pp?: string;
  
    number_objects?: string;
  
    number_complaints?: string;
  
    consturctive?: string;
  
    card_additional_fields?: string;
  
    characteristics_additional_fields?: string;
  
    rest_additional_fields?: string;

    status1?: string;
  
    status2?: string;
  
    cadastr_number?: number;
  
    latitude?: number;
  
    longtitude?: number;
  }